requirements.txt
mcp
fastapi
uvicorn
requests
google-generativeai